[indexer] Check for runtime selectors for static displacement
